requirements.txt
Django==4.2.20
djangorestframework==3.16.0
gunicorn
pymysql==1.1.0
